api/requirements.txt
# Core
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.6

# Financial Core
yfinance==0.2.37
pandas==2.2.1
numpy==1.26.4
pandas-ta==0.3.14b0
pyportfolioopt==1.5.5
quantstats==0.0.62
riskfolio-lib==0.1.8

# Advanced Analytics
tensorflow==2.15.0
prophet==1.1.5
statsmodels==0.14.1

# News & Sentiment
newspaper3k==0.2.8
nltk==3.8.1
vaderSentiment==3.3.2

# Visualization
plotly==5.18.0
matplotlib==3.8.3
seaborn==0.13.2

# Quantitative Finance
pyblackscholes==0.7.1
quantlib-python==1.31
zipline-reloaded==3.0.0
pyfolio==0.9.2
pytorch-forecasting==1.0.0

# Alternative Data
twitter-scraper==2.5.1
sec-edgar-downloader==4.0.5
cryptocompare==0.9.1

# Execution Infrastructure
ccxt==4.2.85
alpaca-trade-api==3.3.0
ibapi==10.24.2

# Blockchain Integration
web3==6.11.1
coinmetrics-api-client==2024.1

# Advanced Visualization
dash==2.14.2
streamlit==1.32.0
